+++ app/src/components/Componentes Detalles/Inscripciones.js
@@ -1,0 +1,14 @@
+import React from "react";
+
+const Inscripciones = () => {
+  return (
+    <div>
+      <h2 className="text-3xl font-bold mb-4">Inscripciones</h2>
+      <p className="text-gray-300">
+        Actualmente, hay 78 jugadores inscritos en este torneo. ¡No pierdas la oportunidad de unirte y demostrar tu habilidad!
+      </p>
+    </div>
+  );
+};
+
+export default Inscripciones;
